requirements.txt
# TS-BOSS runtime dependencies
# Python: 3.12.x

numpy==2.3.4
pandas==2.3.3
scipy==1.16.3
matplotlib==3.10.7
networkx==3.5
scikit-learn==1.7.2

# Causal discovery stack
tigramite==5.2.9.4
jpype1==1.6.0
dcor==0.6
numba==0.62.1

# Notebook execution (optional but recommended)
ipykernel==7.1.0
jupyter_client==8.6.3
jupyter_core==5.9.1
